Reorganized `org.cirdles.topsoil.data` as `org.cirdles.topsoil.dataset`
Reorganized `org.cirdles.topsoil.data` as `org.cirdles.topsoil.dataset`
by renaming the package and adding subpackages for entries and fields.

Also corrected a licensing issue.
