+++ src/main/java/org/cirdles/topsoil/app/chart/VariableBindingControl.java
@@ -23,7 +23,7 @@ import javafx.util.StringConverter;
 import org.cirdles.javafx.CustomHBox;
 import org.cirdles.topsoil.chart.Variable;
 import org.cirdles.topsoil.chart.VariableFormat;
-import org.cirdles.topsoil.data.Field;
+import org.cirdles.topsoil.dataset.field.Field;
 
 /**
  *
