+++ src/main/java/org/cirdles/topsoil/app/chart/VariableBindingDialog.java
@@ -22,7 +22,7 @@ import javafx.scene.control.Dialog;
 import org.cirdles.topsoil.chart.Chart;
 import org.cirdles.topsoil.chart.Variable;
 import org.cirdles.topsoil.chart.VariableContext;
-import org.cirdles.topsoil.data.Dataset;
+import org.cirdles.topsoil.dataset.Dataset;
 
 public class VariableBindingDialog extends Dialog<VariableContext> {
