+++ src/main/java/org/cirdles/topsoil/chart/VariableBinding.java
@@ -15,8 +15,8 @@
  */
 package org.cirdles.topsoil.chart;
 
-import org.cirdles.topsoil.data.Entry;
-import org.cirdles.topsoil.data.Field;
+import org.cirdles.topsoil.dataset.entry.Entry;
+import org.cirdles.topsoil.dataset.field.Field;
 
 /**
  *
