+++ src/main/java/org/cirdles/topsoil/dataset/DatasetManager.java
@@ -13,19 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.data;
+package org.cirdles.topsoil.dataset;
 
-import javafx.util.StringConverter;
+import java.util.List;
 
 /**
  *
  * @author John Zeringue
- * @param <T>
  */
-public interface Field<T> {
+public interface DatasetManager {
     
-    public String getName();
+    public List<Dataset> getDatasets();
     
-    public StringConverter<T> getStringConverter();
+    public boolean isOpen(Dataset dataset);
+    
+    public boolean isClosed(Dataset dataset);
+    
+    public void open(Dataset dataset);
+    
+    public void close(Dataset dataset);
     
 }
