+++ src/main/java/org/cirdles/topsoil/dataset/field/BaseField.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 CIRDLES.
+ * Copyright 2014 CIRDLES.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,24 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.data;
-
-import java.util.List;
+package org.cirdles.topsoil.dataset.field;
 
 /**
  *
- * @author John Zeringue
+ * @author CIRDLES
+ * @param <T>
  */
-public interface DatasetManager {
-    
-    public List<Dataset> getDatasets();
-    
-    public boolean isOpen(Dataset dataset);
-    
-    public boolean isClosed(Dataset dataset);
-    
-    public void open(Dataset dataset);
-    
-    public void close(Dataset dataset);
-    
+public abstract class BaseField<T> implements Field<T> {
+
+    private final String name;
+
+    public BaseField(String name) {
+        this.name = name;
+    }
+
+    /**
+     * @return the name
+     */
+    @Override
+    public String getName() {
+        return name;
+    }
 }
