+++ src/main/java/org/cirdles/topsoil/dataset/field/TextField.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 CIRDLES.
+ * Copyright 2014 zeringuej.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,32 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.data;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
+package org.cirdles.topsoil.dataset.field;
+
+import javafx.util.StringConverter;
 
 /**
  *
- * @author John Zeringue
+ * @author zeringuej
  */
-public class SimpleEntry implements Entry {
-
-    private final Map<Field, Object> fieldsToValues;
-
-    public SimpleEntry() {
-        fieldsToValues = new HashMap<>();
+public class TextField extends BaseField<String> {
+    
+    public TextField(String name) {
+        super(name);
     }
 
     @Override
-    public <T> Optional<T> get(Field<? extends T> field) {
-        return Optional.ofNullable((T) fieldsToValues.get(field));
-    }
+    public StringConverter<String> getStringConverter() {
+        return new StringConverter<String>() {
 
-    @Override
-    public <T> void set(Field<? super T> field, T value) {
-        fieldsToValues.put(field, value);
+            @Override
+            public String toString(String string) {
+                return string;
+            }
+
+            @Override
+            public String fromString(String string) {
+                return string;
+            }
+        };
     }
     
 }
