+++ src/test/java/org/cirdles/topsoil/app/chart/VariableBindingControlTest.java
@@ -22,8 +22,8 @@ import javafx.scene.control.ChoiceBox;
 import javafx.stage.Stage;
 import org.cirdles.topsoil.chart.IndependentVariable;
 import org.cirdles.topsoil.chart.Variable;
-import org.cirdles.topsoil.data.Field;
-import org.cirdles.topsoil.data.NumberField;
+import org.cirdles.topsoil.dataset.field.Field;
+import org.cirdles.topsoil.dataset.field.NumberField;
 import org.junit.Test;
 import static org.testfx.api.FxAssert.verifyThat;
 import org.testfx.framework.junit.ApplicationTest;
