+++ src/test/java/org/cirdles/topsoil/app/utils/TSVDatasetWriterTest.java
@@ -18,12 +18,12 @@ package org.cirdles.topsoil.app.utils;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import org.cirdles.topsoil.data.Dataset;
-import org.cirdles.topsoil.data.Entry;
-import org.cirdles.topsoil.data.Field;
-import org.cirdles.topsoil.data.NumberField;
-import org.cirdles.topsoil.data.SimpleDataset;
-import org.cirdles.topsoil.data.SimpleEntry;
+import org.cirdles.topsoil.dataset.Dataset;
+import org.cirdles.topsoil.dataset.entry.Entry;
+import org.cirdles.topsoil.dataset.field.Field;
+import org.cirdles.topsoil.dataset.field.NumberField;
+import org.cirdles.topsoil.dataset.SimpleDataset;
+import org.cirdles.topsoil.dataset.entry.SimpleEntry;
 import org.junit.Test;
 import static org.junit.Assert.*;
 import org.junit.Before;
